# plans/openapi-worker.yaml
openapi: 3.0.3
info:
  title: SpecTrace Worker API
  version: 1.0.0
  description: |-
    The v1 contract the Cloudflare Worker implements. Every path under `/api/v1/` survives from the Django server with its schemas intact, minus the Linear and GitHub webhook integrations. Five endpoints are new: the CLI pushes the requirement tree (`PUT /api/v1/specs/`), impact reports (`POST /api/v1/results/impact/`), and drift reports (`POST /api/v1/results/drift/`); `spectrace lore sync` drains the task outbox (`GET /api/v1/tasks/outcomes`, `POST /api/v1/tasks/outcomes/ack`). `GET /api/v1/specs/impact/` and `GET /api/v1/specs/drift/` serve the last pushed report instead of computing one; their response shapes stay.

    The Worker enforces the API key on every route: send it in the `X-API-Key` header. A missing or wrong key returns 401 with an `ErrorEnvelope`.

    Two error shapes exist. Routes that moved from `api.py` keep the legacy `ErrorResponse` (`{"success": false, "error": "..."}`); every other route returns `ErrorEnvelope` (`{"error": {"code", "message", "details"}}`). Success bodies on those routes wrap the payload as `{"data": ..., "meta": ...}` with `meta` present only when the route paginates.
servers:
- url: /
security:
- apiKey: []
...
paths:
...
  /api/v1/integrations/slo/status/:
    get:
      tags:
      - SLO
      summary: List SLO status
      operationId: api-v1-integrations-slo-status_get
      responses:
        '200':
          description: All SLOs and their linked requirements
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SLOStatusSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - SLO
      summary: Update SLO status
      description: Update SLO status from observability platform.
      operationId: api-v1-integrations-slo-status_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SLOStatusRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SLOStatusResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': *id001
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: The `SPECTRACE_API_KEY` Worker secret. Required on every route.
...
  responses:
    Unauthorized:
      description: Missing or wrong `X-API-Key`; `code` is `unauthorized`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    SLOStatusSummary:
      type: object
      required:
      - name
      - display_name
      - service
      - time_window
      - budgeting_method
      - status
      - requirement_ids
      properties:
        name:
          type: string
        display_name:
          type: string
        service:
          type: string
        target:
          type: number
          nullable: true
        time_window:
          type: string
        budgeting_method:
          type: string
        status:
          type: string
        current_value:
          type: number
          nullable: true
        error_budget_remaining:
          type: number
          nullable: true
        last_updated:
          type: string
          format: date-time
          nullable: true
        requirement_ids:
          type: array
          items:
            type: string
...
    SLOStatusItem:
      title: SLOStatusItem
      description: Individual SLO status update.
      type: object
      properties:
        name:
          type: string
          maxLength: 500
        status:
          type: string
          maxLength: 20
        current_value:
          default: null
          type: number
          nullable: true
        error_budget_remaining:
          default: null
          type: number
          nullable: true
      required:
      - name
      - status
    SLOStatusRequest:
      title: SLOStatusRequest
      description: Request body for POST /api/v1/integrations/slo/status/.
      type: object
      properties:
        slos:
          type: array
          items:
            $ref: '#/components/schemas/SLOStatusItem'
          maxItems: 200
        update_verification_status:
          type: boolean
          default: false
      required:
      - slos
    SLOStatusResponse:
      title: SLOStatusResponse
      description: Response for POST /api/v1/integrations/slo/status/.
      type: object
      properties:
        success:
          type: boolean
        updated:
          type: integer
        not_found:
          type: integer
        requirement_status:
          type: object
          additionalProperties:
            type: integer
      required:
      - success
      - updated
      - not_found
      - requirement_status
    ErrorResponse:
      title: ErrorResponse
      description: Standard error response.
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          default: ''
      required: []